Refactor manual aws_alias_cloudfront and add integration tests
