+++ manual_scans/aws/aws_alias_cloudfront_s3.py
@@ -9,16 +9,12 @@ from utils.utils_print import my_print
 from utils.utils_print import print_list
 
 
-vulnerable_domains = []
-missing_resources = []
-
-
 def vulnerable_alias_cloudfront_s3(domain_name):
 
     try:
         response = requests.get(f"https://{domain_name}", timeout=1)
 
-        if response.status_code == 404 and "Code: NoSuchBucket" in response.text:
+        if response.status_code == 404 and "<Code>NotFound</Code>" in response.text:
             return True
 
     except (requests.exceptions.ConnectionError, requests.exceptions.ReadTimeout):
@@ -28,6 +24,8 @@ def vulnerable_alias_cloudfront_s3(domain_name):
 
 
 def route53():
+    vulnerable_domains = []
+    missing_resources = []
 
     print("Searching for Route53 hosted zones")
 
@@ -60,12 +58,13 @@ def route53():
                 else:
                     my_print(f"{str(i)}. {record['Name']}", "SECURE")
 
+    return vulnerable_domains, missing_resources
 
-if __name__ == "__main__":
 
+def main():
     parser = argparse.ArgumentParser(description="Prevent Subdomain Takeover")
 
-    route53()
+    vulnerable_domains, missing_resources = route53()
 
     count = len(vulnerable_domains)
     my_print(f"\nTotal Vulnerable Domains Found: {str(count)}", "INFOB")
@@ -76,3 +75,7 @@ if __name__ == "__main__":
 
         my_print("\nCloudFront distributions with missing S3 origin: ", "INFOB")
         print_list(missing_resources, "OUTPUT_WS")
+
+
+if __name__ == "__main__":
+    main()
